third_party/WebKit/LayoutTests/virtual/color_space/fast/canvas/color-space/canvas-drawImage-rec2020.html

<!DOCTYPE HTML>
<script src="../../../../../resources/testharness.js"></script>
<script src="../../../../../resources/testharnessreport.js"></script>
<script>

var rec2020TransparentRed = [0.294694, 0.014558, 0.024679, 0.501961];
var rec2020TransparentGreen = [0.038849, 0.302285, 0.115265, 0.501961];
var rec2020TransparentBlue = [0.016155, 0.032855, 0.209754, 0.501961];
var rec2020TransparentBlack = [0.010960, 0.010960, 0.010960, 0.501961];

function testPixels(ctx, tests)
{
    var actual, expected, tolerance = 0.09;
    console.log(tests.length);
    for (var i = 0; i < tests.length; i++) {
      actual = ctx.getImageData(tests[i].x, tests[i].y, 1, 1).dataUnion;
      expected = tests[i].color;
      assert_true(actual.length === expected.length);
      for (var j = 0; j < actual.length; j++)
        assert_approx_equals(actual[j], expected[j], tolerance);
    }
}

function drawSRGBImageOnRec2020Canvas(source)
{
  var canvas = document.createElement('canvas');
  canvas.width = 20;
  canvas.height = 20;
  var ctx = canvas.getContext('2d',
      {colorSpace: 'rec2020', pixelFormat:'float16'});
  ctx.drawImage(source, 0, 0);
  var tests = [{x: 5, y: 5, color: rec2020TransparentRed},
               {x: 15, y: 5, color: rec2020TransparentGreen},
               {x: 5, y: 15, color: rec2020TransparentBlue},
               {x: 15, y: 15, color: rec2020TransparentBlack}];
  testPixels(ctx, tests);
}

promise_test(function() {
  return new Promise((resolve, reject) => {
    var image = new Image();
    image.onload = function() {
      resolve(image);
    }
    image.src = 'resources/pattern-semitransparent-srgb.png'
  }).then(drawSRGBImageOnRec2020Canvas);
}, 'Draw SRGB image on a Rec2020 canvas and read back the Rec2020 pixels.');

// TODO(zakerinasab): crbug.com/713867
// Add at least two more tests to load pattern-semitransparent-p3d65.png and
// pattern-semitransparent-rec2020.png and draw them on a Rec2020 canvas.

</script>
